+++ src/rx-collection.js
@@ -32,7 +32,8 @@ export class RxCollection {
         migrationStrategies = {},
         methods = {},
         attachments = {},
-        options = {}
+        options = {},
+        statics = {}
     ) {
         this._isInMemory = false;
         this.destroyed = false;
@@ -44,6 +45,7 @@ export class RxCollection {
         this._methods = methods; // orm of documents
         this._attachments = attachments; // orm of attachments
         this.options = options;
+        this._statics = statics;
         this._atomicUpsertQueues = {};
 
         this._docCache = DocCache.create();
@@ -776,7 +778,8 @@ export async function create({
         migrationStrategies,
         methods,
         attachments,
-        options
+        options,
+        statics
     );
     await collection.prepare();